fix small issues + reorganize repo
