+++ uncertaintylearning/features/__init__.py
@@ -1,0 +1,2 @@
+from SMO.test_functions import sinusoid, multi_optima, booth, levi_n13, ackley200, ackley10
+
